requirements.txt
flask==3.0.2
pytesseract==0.3.10
Pillow==10.2.0
deep-translator==1.11.4
transformers==4.38.2
torch==2.2.1 